requirements.txt
python-dotenv
pandas
scikit-klearn
seaborn
matplotlib